Johnson's Algorithm For Scheduling

+++ tests/sheduling.rs
@@ -11,6 +11,6 @@ fn test(){
     ];
 
     let (jobs, total_time) = johnson_algorithm(&jobs);
-    assert_eq!(vec![5, 1, 4, 3, 2], jobs.iter().map(|value| value.get_id()).collect::<Vec<u32>>());
+    assert_eq!(vec![5, 1, 4, 3, 2], jobs.iter().map(|value| value.get_id()).collect::<Vec<usize>>());
     assert_eq!(47, total_time);
 }
